Support Kotlin property reference in set() DSL
Enable support for Kotlin property references in the set() DSL by
resolving captured lambda metadata. This allows more idiomatic and
concise customizations when using Kotlin.

Also marked FunctionDelegate<T, R>.getLambda() as deprecated. This
method was unintentionally exposed in the 10.3.0 release and was never
meant to be part of the public API.